% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plotGens}
\alias{plotGens}
\title{Plot of cases by generations}
\usage{
plotGens(out = NULL, saveFile = F)
}
\arguments{
\item{saveFile}{}
}
\value{

}
\description{
Plot of cases by generations
}
\examples{
plotGens(saveFile = T)
}
