% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plotOutput}
\alias{plotOutput}
\title{Plot summary, trend, and seasonality}
\usage{
plotOutput(val, out, saveFile = F)
}
\arguments{
\item{saveFile}{}
}
\value{

}
\description{
Plot summary, trend, and seasonality
}
\examples{
plotOutput(val, out, saveFile = T)
}
