% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{stanBstsFit}
\alias{stanBstsFit}
\title{Fit a model with RStan}
\usage{
stanBstsFit(num = NULL, lastday = NULL)
}
\arguments{
\item{num}{}
}
\value{
out
}
\description{
Fit a model with RStan
}
\examples{
fit <- stanBstsFit(100)
}
