% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{textOutput}
\alias{textOutput}
\title{Texts for reporting on Twitter}
\usage{
textOutput(val, num)
}
\arguments{
\item{num}{}
}
\value{

}
\description{
Texts for reporting on Twitter
}
\examples{
textOutput(val, 100)
}
