% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{textOutputs}
\alias{textOutputs}
\title{Texts for reporting on Twitter}
\usage{
textOutputs(val, num)
}
\arguments{
\item{num}{}
}
\value{

}
\description{
Texts for reporting on Twitter
}
\examples{
textOutputs(val, 100)
}
